Add files via upload
This is the first version of the yelp camp website. It includes a basic styling, forms and couple of pages where the users can add a new campground, view all the campgrounds. It's built with the help of HTML, CSS, EJS, node, Express. The next version will have better styling and connection to mongoose DB.